ALL CREDIT TO OSTEP.ORG
I simply working through book code, and wanting to save to keep track of
where I am in the learning process. I am not creating anything. If I
modify the code in any manner it will be noted. Otherwise all code is
property of ostep.org. Not mine. Emphasis; not mine.